# openapi.yaml
openapi: 3.0.3
info:
  title: Hadi Hep Beraber - Okey Game API
  description: API for Turkish Okey game platform
  version: 1.0.0
servers:
  - url: /api
    description: Next.js API routes
...
paths:
...
  /games:
    get:
      tags: [games]
      summary: List available games to join
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [WAITING]
            default: WAITING
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GameLobby'

    post:
      tags: [games]
      summary: Create a new game
      security:
        - session: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mode:
                  type: string
                  enum: [regular, okey101]
                  default: regular
                maxPlayers:
                  type: integer
                  enum: [2, 3, 4]
                  default: 4
                isPrivate:
                  type: boolean
                  default: false
                turnTimeLimit:
                  type: integer
                  default: 30
                fillWithAI:
                  type: boolean
                  default: true
                  description: Auto-fill with AI players and start immediately
                rules:
                  $ref: '#/components/schemas/GameRules'
                stake:
                  type: object
                  description: Entry fee and pot (optional, for chip games)
                  properties:
                    entryFee:
                      type: integer
                      description: Chips required to join
                    potDistribution:
                      type: string
                      enum: [winner_takes_all, proportional]
                      default: winner_takes_all
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '400':
          description: Not enough chips for stake
...
  /games/join-by-code:
    post:
      tags: [games]
      summary: Join a private game by room code
      security:
        - session: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [roomCode]
              properties:
                roomCode:
                  type: string
                  example: "ABC123"
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '404':
          description: Invalid room code

  # ============================================
  # GAME ACTIONS (REST fallback, prefer Socket.io)
  # ============================================
...
  /games/{gameId}/validate-hand:
    post:
      tags: [game-actions]
      summary: Validate hand without finishing
      description: |
        Check if current hand is a valid winning hand.
        Useful for client-side validation, AI testing, and cheat detection.
      security:
        - session: []
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                discardTileId:
                  type: string
                  description: Tile to exclude when validating (simulates discard)
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  isValid:
                    type: boolean
                  groups:
                    type: array
                    items:
                      $ref: '#/components/schemas/TileGroup'
                  remainingTiles:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tile'
                  errorMessage:
                    type: string

  # ============================================
  # GAME EVENTS & REPLAY
  # ============================================
...
  /shop/chips/purchase:
    post:
      tags: [shop]
      summary: Purchase chips
      security:
        - session: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [packageId]
              properties:
                packageId:
                  type: string
      responses:
        '200':
          description: Returns payment URL or confirms purchase
...
  /shop/cosmetics/{cosmeticId}/purchase:
    post:
      tags: [shop]
      summary: Purchase cosmetic with chips
      security:
        - session: []
      parameters:
        - name: cosmeticId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Cosmetic purchased
        '400':
          description: Not enough chips or already owned

  # ============================================
  # LEADERBOARD
  # ============================================
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: next-auth.session-token
  schemas:
...
    UserPublic:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        image:
          type: string
        rating:
          type: integer
        isVip:
          type: boolean
...
    Tile:
      type: object
      properties:
        id:
          type: string
        number:
          type: integer
          minimum: 1
          maximum: 13
        color:
          type: string
          enum: [red, yellow, blue, black]
        tileType:
          type: string
          enum: [normal, fake_okey]
          description: |
            - normal: Regular numbered tile
            - fake_okey: Sahte okey (joker tile) - represents the real okey's value only, NOT a wildcard
            Note: Whether a tile IS the real okey is determined dynamically (indicator + 1)
    TileGroup:
      type: object
      properties:
        type:
          type: string
          enum: [set, run]
          description: |
            - set (per): 3-4 tiles, same number, different colors
            - run (seri): 3+ consecutive tiles, same color
        tiles:
          type: array
          items:
            $ref: '#/components/schemas/Tile'
    GameRules:
      type: object
      description: Rule configuration for a game variant
      properties:
        variant:
          type: string
          enum: [okey, okey101]
          default: okey
        openingRequired:
          type: boolean
          default: false
          description: Must open (show valid groups) before finishing
        openingMinScore:
          type: integer
          default: 0
          description: Minimum score to open (101 variant)
        allowCift:
          type: boolean
          default: true
          description: Allow çift (7 pairs) win
        timeoutPolicy:
          type: string
          enum: [auto_discard_random, auto_discard_last_drawn, skip_turn]
          default: auto_discard_random
          description: What happens when turn times out
...
    Game:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
          enum: [WAITING, STARTING, PLAYING, FINISHED, CANCELLED]
        mode:
          type: string
          enum: [regular, okey101]
        maxPlayers:
          type: integer
        isPrivate:
          type: boolean
        roomCode:
          type: string
        players:
          type: array
          items:
            $ref: '#/components/schemas/GamePlayer'
        currentTurn:
          type: integer
        turnPhase:
          type: string
          enum: [draw, discard]
        indicatorTile:
          $ref: '#/components/schemas/Tile'
        okeyTile:
          $ref: '#/components/schemas/Tile'
        discardPileTop:
          $ref: '#/components/schemas/Tile'
        tileBagCount:
          type: integer
        turnTimeLimit:
          type: integer
        turnStartedAt:
          type: string
          format: date-time
        rules:
          $ref: '#/components/schemas/GameRules'
        stake:
          type: object
          properties:
            entryFee:
              type: integer
            pot:
              type: integer
            potDistribution:
              type: string
        winnerId:
          type: string
        finishType:
          type: string
          enum: [normal, cift, el]
          description: |
            - normal: Standard finish
            - cift: 7 pairs (2x score in 101)
            - el: First turn win (no draw)
    GameLobby:
      type: object
      description: Game info shown in lobby (no sensitive data)
      properties:
        id:
          type: string
        mode:
          type: string
        maxPlayers:
          type: integer
        currentPlayers:
          type: integer
        host:
          $ref: '#/components/schemas/UserPublic'
    GamePlayer:
      type: object
      properties:
        id:
          type: string
        position:
          type: integer
        name:
          type: string
        avatar:
          type: string
        isAI:
          type: boolean
        isConnected:
          type: boolean
        isReady:
          type: boolean
        tileCount:
          type: integer
          description: Number of tiles (tiles themselves hidden for opponents)
        tiles:
          type: array
          items:
            $ref: '#/components/schemas/Tile'
          description: Only included for the requesting player
...